Revert "Bring over pdf fill dependency to the pension module (#17573)"
This reverts commit 3d931acdaabf1cd6487b542606577e8dbe13986f.
