+++ modules/pensions/app/models/pensions/saved_claim.rb
@@ -1,6 +1,5 @@
 # frozen_string_literal: true
 
-# Pensions Module
 module Pensions
   ##
   # Pension 21P-527EZ Active::Record
