+++ modules/pensions/documentation/adr/0007-pdf-filler-library.md
@@ -20,12 +20,8 @@ In the main app there is a lib called `pdf_filler` and this contains the informa
 
 ## Decision
 
-The PDF Filler presents its own set of challenges to bring together as much of these files without changing the core library too much. We looked into bringing over the files with minimal impact to the rest of the team projects during this follow up PR. We decided to modify the core files to create a register function. We then isolated any pathing related concerns and made modifications. This allowed us to be able to isolate and bring over the Pension modules specific set of tests and fixture files.
+The PDF Filler presents its own set of challenges to bring together as much of these files without changing the core library too much.
 
 ## Consequences
 
-The core ownership has moved to the `lib/pdf_fill` and `spec/pdf_fill` folders under the pension module. This allows us to isolate the work PDF work being done for the pension team to the module folder. As a result the team has less concerns when working within the scope of modifying the pension form code. 
-
-There is an registration function in the `lib/pdf_fill/filler.rb` that requires a `Style/MutableConstant` ignore so we can have a place to create a registration for our forms. If there is a better way to do this practice we could improve the readability here.
-
-One challenge that is worth noting is that the naming was decided because of copying the files over from the root application. We want to challenge ourself here to think about the proper acceptance criteria for these incoming file paths and namespaces, but also think that it may be better to address improvements and path name improvements in future work.
+Look into bringing over the files with minimal impact to the rest of the team projects. Perhaps as a follow up PR.
