--- modules/pensions/lib/pensions/version.rb
@@ -1,13 +1,0 @@
-# frozen_string_literal: true
-
-module Pensions
-  ##
-  # The module path
-  #
-  MODULE_PATH = 'modules/pensions'
-
-  ##
-  # The module version
-  #
-  VERSION = '0.1.0'
-end
