+++ spec/lib/pdf_fill/filler_spec.rb
@@ -2,7 +2,6 @@
 
 require 'rails_helper'
 require 'pdf_fill/filler'
-require 'lib/pdf_fill/fill_form_examples'
 
 describe PdfFill::Filler, type: :model do
   include SchemaMatchers
@@ -48,12 +47,16 @@ describe PdfFill::Filler, type: :model do
     end
   end
 
-  # see `fill_form_examples.rb` for documentation about options
-  describe '#fill_form' do
+  describe '#fill_form', run_at: '2017-07-25 00:00:00 -0400' do
     [
       {
         form_id: '21P-530V2',
         factory: :burial_claim_v2
+      },
+      {
+        form_id: '21P-527EZ',
+        factory: :pension_claim,
+        use_vets_json_schema: true
       },
       {
         form_id: '21P-0969',
@@ -63,8 +66,17 @@ describe PdfFill::Filler, type: :model do
       {
         form_id: '10-10CG',
         factory: :caregivers_assistance_claim,
-        input_data_fixture_dir: 'spec/fixtures/pdf_fill/10-10CG',
-        output_pdf_fixture_dir: 'spec/fixtures/pdf_fill/10-10CG/signed',
+        input_data_fixture_dir: 'pdf_fill/10-10CG',
+        output_pdf_fixture_dir: 'pdf_fill/10-10CG/signed',
+        fill_options: {
+          sign: true
+        }
+      },
+      {
+        form_id: '10-10CG',
+        factory: :caregivers_assistance_claim,
+        input_data_fixture_dir: 'pdf_fill/10-10CG',
+        output_pdf_fixture_dir: 'pdf_fill/10-10CG/signed',
         fill_options: {
           sign: true
         }
@@ -74,7 +86,66 @@ describe PdfFill::Filler, type: :model do
         factory: :dependency_claim
       }
     ].each do |options|
-      it_behaves_like 'a form filler', options
+      form_id, factory = options.values_at(:form_id, :factory)
+
+      context "form #{form_id}" do
+        %w[simple kitchen_sink overflow].each do |type|
+          context "with #{type} test data" do
+            let(:input_data_fixture_dir) { options[:input_data_fixture_dir] || "pdf_fill/#{form_id}" }
+            let(:output_pdf_fixture_dir) { options[:output_pdf_fixture_dir] || "pdf_fill/#{form_id}" }
+            let(:form_data) do
+              return get_fixture("#{input_data_fixture_dir}/#{type}") unless options[:use_vets_json_schema]
+
+              schema = "#{form_id.upcase}-#{type.upcase}"
+              VetsJsonSchema::EXAMPLES.fetch(schema)
+            end
+            let(:saved_claim) do
+              if form_id == '21P-530V2'
+                claim = create(factory)
+                claim.update(form: form_data.to_json)
+                claim
+              else
+                create(factory, form: form_data.to_json)
+              end
+            end
+
+            it 'fills the form correctly' do
+              if type == 'overflow'
+                # pdfs_fields_match? only compares based on filled fields, it doesn't read the extras page
+                the_extras_generator = nil
+
+                expect(described_class).to receive(:combine_extras).once do |old_file_path, extras_generator|
+                  the_extras_generator = extras_generator
+                  old_file_path
+                end
+              end
+
+              file_path = if options[:fill_options]
+                            described_class.fill_form(saved_claim, nil, options[:fill_options])
+                          else
+                            # Should be able to call without any additional arguments
+                            described_class.fill_form(saved_claim)
+                          end
+
+              if type == 'overflow'
+                extras_path = the_extras_generator.generate
+
+                expect(
+                  FileUtils.compare_file(extras_path, "spec/fixtures/#{output_pdf_fixture_dir}/overflow_extras.pdf")
+                ).to eq(true)
+
+                File.delete(extras_path)
+              end
+
+              expect(
+                pdfs_fields_match?(file_path, "spec/fixtures/#{output_pdf_fixture_dir}/#{type}.pdf")
+              ).to eq(true)
+
+              File.delete(file_path)
+            end
+          end
+        end
+      end
     end
   end
